chore: allow calculating prices without explicit fields
